fix: Move exceptions module to rossum_api package
It's not importable like this if user only installs the rossum-api package.
