+++ rossum_api/__init__.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 
-from exceptions import APIClientError
 from rossum_api.clients.external_async_client import AsyncRossumAPIClient
 from rossum_api.clients.external_sync_client import SyncRossumAPIClient
+from rossum_api.exceptions import APIClientError
 
 __version__ = "2.0.0"
