+++ rossum_api/clients/internal_async_client.py
@@ -7,7 +7,6 @@ import typing
 import httpx
 import tenacity
 
-from exceptions import APIClientError
 from rossum_api.domain_logic.annotations import (
     build_export_query_params,
     get_http_method_for_annotation_export,
@@ -23,6 +22,7 @@ from rossum_api.domain_logic.urls import (
     build_upload_url,
     build_url,
 )
+from rossum_api.exceptions import APIClientError
 from rossum_api.utils import enforce_domain
 
 if typing.TYPE_CHECKING:
